requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
black==25.1.0
bleach==6.1.0
cachetools==5.3.3
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cubinlinker-cu11==0.3.0.post2
cuda-python==11.8.3
cudf-cu11==24.4.0
cuml-cu11==24.4.0
cupy-cuda11x==13.1.0
cycler==0.12.1
dask==2024.1.1
dask-cuda==24.4.0
dask-cudf-cu11==24.4.0
dask-expr==0.4.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distributed==2024.1.1
docker-pycreds==0.4.0
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.19
fastjsonschema==2.20.0
fastrlock==0.8.2
filelock==3.13.1
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.73.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
imagecodecs==2024.1.1
importlib_metadata==7.1.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.24.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
lightning-utilities==0.11.2
littleutils==0.2.2
llvmlite==0.42.0
locket==1.0.0
lxml==5.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl-fft 
mkl-random 
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.8
mypy_extensions==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
notebook_shim==0.2.4
numba==0.59.1
numcodecs==0.12.1
numpy 
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
nvtx==0.2.10
ogb==1.3.6
omegaconf==2.3.0
outdated==0.2.2
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.1
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
portalocker==2.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptxcompiler-cu11==0.8.1.post1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pycparser==2.22
Pygments==2.17.2
pylibraft-cu11==24.4.0
pynvml==11.4.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
raft-dask-cu11==24.4.0
rapids-dask-dependency==24.4.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rmm-cu11==24.4.0
rpds-py==0.19.1
scikit-learn==1.4.2
scipy 
Send2Trash==1.8.3
sentry-sdk==2.0.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
submitit==1.5.1
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.4.24
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
torchmetrics==1.3.2
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
treelite==4.1.2
triton==2.2.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
ucx-py-cu11==0.37.0
uri-template==1.3.0
urllib3==1.26.13
wandb==0.16.6
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wilds==2.0.0
xformers==0.0.24+cu118
yacs==0.1.8
zarr==2.17.2
zict==3.0.0
zipp==3.18.1
